Adds L2 Global Exit Root Manager
Implements a Global Exit Root Manager contract for L2 chains.

This commit introduces a new contract, `GlobalExitRootManagerL2SovereignChain`, to manage exit roots for Sovereign L2 chains. It also updates the deployment scripts for both L1 and L2 to deploy and initialize this new contract, and register rollup on L1. This enables the management of global exit roots specifically for sovereign chains, adding flexibility and control over the cross-chain communication process.