Waits for contracts before starting AggKit
Replaces the simple delay with a contract verification step.
This ensures that AggKit only starts after the necessary
contracts are deployed and accessible, preventing startup
errors. It checks for the `POLYGON_ROLLUP_MANAGER_L1`
contract's code on Anvil L1 using `eth_getCode`. If the code
exists, it proceeds. If not, it retries for up to 150
seconds.

Also updates claims.json with appropriate MainnetExitRoot value
and renames aggkit image name in docker compose.