Aggkit claim (#59)
* include claimsponsor logic from aggkit and initial cli layout for this service

* fix metadata in the claim body

* Waits for contracts before starting AggKit

Replaces the simple delay with a contract verification step.
This ensures that AggKit only starts after the necessary
contracts are deployed and accessible, preventing startup
errors. It checks for the `POLYGON_ROLLUP_MANAGER_L1`
contract's code on Anvil L1 using `eth_getCode`. If the code
exists, it proceeds. If not, it retries for up to 150
seconds.

Also updates claims.json with appropriate MainnetExitRoot value
and renames aggkit image name in docker compose.

* Updates sponsor-claim command description

Refines the `long_about` description for the `sponsor-claim` command.

This provides a clearer and more concise explanation of the command's
functionality and usage within the CLI help text.

* changed aggkit image

---------

Co-authored-by: cicr99 <kr1000a@gmail.com>