Deploys and registers L2 and L3 rollups
Adds deployment and registration logic for L2 and L3 rollups within the L1 deployment script.

This change facilitates the setup of a multi-layered scaling solution by deploying and registering separate PolygonZkEVM instances for L2 and L3.

It also updates the AggKit docker-compose files to use the latest image and include the new rollup addresses in the environment variables. This ensures that AggKit is aware of the new L2 and L3 rollups and can interact with them correctly.

Updates the test suite with correct values for testing the L1/L2 bridge.